+++ src/libs/headers/drivers/kbd.h
@@ -16,14 +16,15 @@
 
 #include <stdint.h>
 #include <stddef.h>
+#include "types.h"
 
-// Structure of a keyboard packet 
+// Structure of a keyboard packet
 struct keyPacket {
   uint32_t scancode;
   uint32_t characterForm;   // Includes the scancode form as a character from keyboard(available only if it is a character) otherwise 0
   uint32_t pressedStatus;   // The status if the key is pressed(1) or released(0)
   uint8_t shortcutStatus;   // If the keypress was a shortcut(e.g. pressed while the control key is pressed)
-  uint8_t shortcutKey;      // The that was pressed together with it(e.g. ctrl) 
+  uint8_t shortcutKey;      // The that was pressed together with it(e.g. ctrl)
   uint8_t capslockStatus;   // Status of capslock - off(0), on(1)
   uint8_t specialkeyStatus; // Tells if the key is spacial and is not used in the typical keyboard function(e.g. home)
 };
@@ -34,14 +35,14 @@ typedef void (*callroutine_t)(uint8_t, uint8_t, uint32_t);
 extern int keyboardInit(uint8_t mode);
 
 // Changes or returns current mode
-extern uint8_t keyboardMode(int command);
+extern int keyboardMode(int command);
 
 // Sets a function pointer to be called when keyboard interrupt is sent
 extern int keyboardCallFunc(callroutine_t callroutine_func);
 
 // Reads amount from keyboard to buf
-extern int keyboardRead(void* buf, size_t amount);
+extern ssize_t keyboardRead(void* buf, size_t amount);
 
 // Writes amount from buf to keyboard(always fails)
-extern int keyboardWrite(void* buf, size_t amount);
+extern ssize_t keyboardWrite(void* buf, size_t amount);
 #endif
